naive, fake credential issuance
Implements a trivial, naive, fake credential issuance flow in which:

- a person can be instructed to obtain credentials from some issuer
- it then makes an RPC to the issuer requesting credentials
- the issuer then makes an RPC to the person's wallet to store the
  issued credential

The intent of having the issuer talk to the wallet is to simulate a
person plugging their phone into an appliance at a police station, or
waving it near an NFC/Bluetooth enabled appliance. Eventually, we'll use
this link so that the wallet and issuer can verify trust and evaluate
policy against each other.

For now, there's no cryptography involved here, and no policy checks
except that I defined two issuers and restricted which credential each
is allowed to issue.