BBS sign issued credentals
Issuer will now generate a deterministic keypair at startup (this is a
hack because I haven't yet worked out anything for trust or key
distribution) and use it to sign issued credentials. Each credential
type gets broken up into a vector of BBS messages (e.g. a distinct
message for each of a driver's licence's name, address or organ donor
consent fields).

The test setup now verifies that the signatures on the credentials in
the wallet are valid.