BBS proof generation
The wallet and person now expose endpoints for generating proofs from
signatures previously obtained from an issuer. We don't yet have a
relying party, but the issuance simulation validates the proof. We prove
possession of a driver's license (without revealing any of the DL's
contents) and the name on the DL (without revealing any other of the
DL's contents).